Migrate prismascans to demonsect
