+++ src/rust/madara/sources/demonsect/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "prismascans"
+name = "demonsect"
 version = "0.1.0"
 edition = "2021"
 publish = false
