+++ src/rust/madara/sources/demonsect/res/filters.json
