+++ src/rust/madara/sources/demonsect/res/settings.json
