--- src/rust/madara/sources/prismascans/res/source.json
@@ -1,10 +1,0 @@
-{
-	"info": {
-		"id": "pt-br.prismascans",
-		"lang": "pt-br",
-		"name": "Prisma Scans",
-		"version": 2,
-		"url": "https://prismacomics.com",
-		"nsfw": 2
-	}
-}
